requirements.txt
pandas>=2.0
python-dotenv>=1.0
requests>=2.28
httpx[http2]>=0.27
beautifulsoup4>=4.12
playwright>=1.40
